refactor introspection logic into hook
