+++ src/screens/database/views/graphql/GraphqlView/index.tsx
@@ -23,8 +23,7 @@ import { parseValue } from "~/util/surrealql";
 import { useIntent } from "~/hooks/url";
 import { useIsLight } from "~/hooks/theme";
 import { useViewEffect } from "~/hooks/view";
-import { buildClientSchema, getIntrospectionQuery, GraphQLSchema } from "graphql";
-import { sleep } from "radash";
+import { useGraphqlIntrospection } from "~/hooks/graphql";
 
 const QueryPaneLazy = memo(QueryPane);
 const VariablesPaneLazy = memo(VariablesPane);
@@ -35,7 +34,6 @@ export function GraphqlView() {
 	const { setGraphqlResponse } = useDatabaseStore.getState();
 
 	const [isEnabled, setEnabled] = useState(false);
-	const [schema, setSchema] = useState<GraphQLSchema | null>(null);
 	const [variablesValid, setVariablesValid] = useState(true);
 	const [queryValid, setQueryValid] = useState(true);
 
@@ -43,6 +41,7 @@ export function GraphqlView() {
 	const isConnected = useIsConnected();
 	const connection = useActiveConnection();
 	const activeView = useConfigStore(state => state.activeView);
+	const [schema, introspectSchema] = useGraphqlIntrospection();
 
 	const isAvailable = GQL_SUPPORTED.has(connection.authentication.protocol);
 	const isSandbox = connection.id === "sandbox";
@@ -77,29 +76,6 @@ export function GraphqlView() {
 	});
 
 	const isValid = queryValid && variablesValid && isEnabled;
-
-	const introspectSchema = async () => {
-		try {
-			// We must wait for the NS/DB to be selected before
-			// starting the introspection process.
-			await sleep(500);
-
-			const query = getIntrospectionQuery();
-			const response = await executeGraphql(query, {});
-
-			if (!response.success) {
-				throw new Error(response.result);
-			}
-
-			const result = JSON.parse(response.result);
-			const schema = buildClientSchema(result.data);
-
-			setSchema(schema);
-		} catch(err: any) {
-			console.warn("Failed to introspect GraphQL schema", err);
-			setSchema(null);
-		}
-	};
 
 	useViewEffect("graphql", () => {
 		if (isAvailable && isConnected) {
